+++ CHANGES.md
@@ -16,6 +16,8 @@ Unreleased
   - `py_gql.schema.InputField`: the resulting object will use `python_name` as a key instead of the exposed field name.
   - `py_gql.schema.Field`: The default resolver will use `python_name` instead of the exposed name when looking the key or attribute in the root object.
 
+- Added `ResolveInfo.get_directive_arguments`
+
 ### Fixed
 
 - Introducing `pyproject.toml` broke Cython integration through build isolation. This is now fixed by always requiring cython as a a build dependency (through `build-system.requires`) and only enabling support behind the `PY_GQL_USE_CYTHON` env variable.